refactor: 메서드 통일화 및 response->dto 변환
