+++ src/main/java/com/sb09/hrbank/mapper/DepartmentMapper.java
@@ -1,6 +1,6 @@
 package com.sb09.hrbank.mapper;
 
-import com.sb09.hrbank.dto.common.DepartmentResponse;
+import com.sb09.hrbank.dto.response.DepartmentDto;
 import com.sb09.hrbank.dto.request.DepartmentCreateRequest;
 import com.sb09.hrbank.entity.Department;
 import org.mapstruct.Mapper;
@@ -10,5 +10,5 @@ public interface DepartmentMapper {
 
   Department toEntity(DepartmentCreateRequest request);
 
-  DepartmentResponse toDto(Department entity, Integer employeeCount);
+  DepartmentDto toDto(Department entity, Integer employeeCount);
 }
